+++ src/main/java/entity/Entity.java
@@ -2,20 +2,20 @@ package entity;
 import java.io.Serializable;
 public abstract class Entity implements Serializable, Cloneable {
 
-    private Integer id;
+    private Long id;
 
     public Entity() {
     }
 
-    public Entity(int id) {
+    public Entity(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
